Fix Black linter
The Black linter was configured with an "exclude" file that was not correctly made - it was excluding anything with "dist" in the path, which included eachdist.py and the opentelemetry-distro directory. By default[1], Black uses the .gitignore file, which is a superset of what was specified in the exclude file, so just use that. Also ran and applied the linter.

[1] https://black.readthedocs.io/en/stable/usage_and_configuration/file_collection_and_discovery.html