Added a random-traffic-generator for the Python Sample App (#31)
*Description of changes:*
Added a random-traffic-generator for the Python Sample App. More details
can be found in the README under
`aws-otel-python-instrumentation/sample-applications/vehicle-dealership-sample-app/random-traffic-generator`.

The traffic generator generates the following traffic:
1. Every minute, sends a single POST request to the VehicleInventoryApp
and sends a single GET request.
2. Every hour, sends a burst of requests: 5 POST requests to the
VehicleInventoryApp and 5 GET requests.
3. Every 5 minutes, sleeps for random amount of time between 30-60
seconds and then sends a GET request to the VehicleInventoryApp with a
random throttle param between 5-20 seconds. The backend reads that
throttle param and simulates throttling for that amount of time before
responding to the request.
4. Every 5 minutes, sleeps for random amount of time between 30-60
seconds and then sends a GET request to the VehicleInventoryApp with an
invalid car id to trigger 404 error.
5. Every 5 minutes, sleeps for random amount of time between 30-60
seconds and then sends a GET request to the ImageServiceApp with a non
existent image name to trigger 500 error due to S3 Error: "An error
occurred (NoSuchKey) when calling the GetObject operation: The specified
key does not exist."

By submitting this pull request, I confirm that you can use, modify,
copy, and redistribute this contribution, under the terms of your
choice.