Set up contract tests
In this commit, we are setting up contract tests similar to those in the
aws-otel-java-instrumentation repo. Initially, we are just commiting the
logic to run the mock collector and a single application that uses the
requests library, but the tests are extensible to other frameworks.

There are a number of small differences between these tests and the Java
contract tests, most of which are language dependent.

One callout is that the building and running of these tests is currently
possible, but not rigourously so. In coming commits, we will make github
workflows to do this, and several build-related configurations/etc. are
likely to change.