Remove `set_http_status` logic from AwsMetricAttributeGenerator
In this commit, we are removing the `set_http_status` logic from AwsMetricAttributeGenerator. This logic was originally in place for parity with Java. In Java, the OTEL instrumentation for AWS SDK does not populate the `http.status_code` attribute for spans, so we had to add logic to extract status code from the exception embedded within the events of the span. However, in Python, we have verified that this is not a problem; `http.status_code` is reliably set and so we do not need this logic. `set_http_status` logic is no longer needed, so this commit seeks to clean up the code by removing it and all code depending on it.