Run contract tests in pr- and main- build workflows (#54)
In this PR, we are adding contract test runs to the pr-build and
main-build, following suit with ADOT Java's
[pr-build](https://github.com/aws-observability/aws-otel-java-instrumentation/blob/61460af22255d5963df17c77a37ef71ff6f652d8/.github/workflows/pr-build.yml#L90)
and
[main-build](https://github.com/aws-observability/aws-otel-java-instrumentation/blob/61460af22255d5963df17c77a37ef71ff6f652d8/.github/workflows/main-build.yml#L184).
Like in Java, we are running these steps after building and pushing
artifacts so we can debug more effectively, if needed. Unlike Java, we
are not separating into separate jobs, as there is no clear benefit to
doing this, as it just incurs overhead of checking out and building
code.

By submitting this pull request, I confirm that you can use, modify,
copy, and redistribute this contribution, under the terms of your
choice.