Fix UnknownRemoteOperation for Database instrumentation. (#50)
Description of changes:
When `"db.operation"` is not provided in span attributes, the
`"aws.remote.operation"` will be `UnknowRemoteOperation`.
This PR updated the `_set_remote_service_and_operation` function to
parsing the `"db.statement"` attribute when `"db.operation"` is missing,
and extract the valid `"aws.remote.operation"` (We setup a set of valid
database related `"aws.remote.operation"` for security concern):
1. Add a json file `configuration/dialect_keywords.json` saving all the
valid keywords. In this file, the sequence of key words matter -> The
keyword with longer word length are placed towards the front of the list
so it will be matched first.
2. We retrieve the first 27 characters from `"db.statement"` to avoid
the case where very large statements, and use regular expression to
match the keyword, it will match the beginning of the string. If the
string's start does not conform to the regular expression, the match
fails.
3. Add unit test for _set_remote_service_and_operation performance where
tested _DB_STATEMENT is/is not present and is/is not valid.
4. Add unit test covering different `"db.statement"` cases:
a. Only 1 valid keywords match
b. More than 1 valid keywords match, we want to pick the longest match
c. More than 1 valid keywords match, but the other keywords is not at
the start of the SpanAttributes.DB_STATEMENT. We want to only pick start
match
d. No valid match
e. Have valid but it is not at the start of SpanAttributes.DB_STATEMENT
f. Have valid keywords, match the longest word
g. Have valid keywords, match with first word
h.Have valid keywords, match with upper case
5.Add unit test for testing keywords sequence in json file:
a.Confirm the keywords are sorted based on descending order of keywords
character length
b.Confirm maximum length of keywords is not longer than
MAX_KEYWORD_LENGTH
6. exclude `configuration/dialect_keywords.json` from codespell check
because the Keywords are fixed for the query, and codespell does not
recognize the comments in json.

Testing
Tested by deploying the code changes to EKS and nodes, where we can see
the actual operation instead of `UnknowRemoteOperation`:

![Screenshot 2024-02-13 at 6 21
34 PM](https://github.com/aws-observability/aws-otel-python-instrumentation/assets/146124015/cd4cb2d5-f069-45c9-8487-45a709ef6e5a)


By submitting this pull request, I confirm that you can use, modify,
copy, and redistribute this contribution, under the terms of your
choice.