Improve build process by moving to script (#63)
In this commit, we are moving the build and install distro steps to a
script to make it easier to run. As well, we are moving
`set-up-contract-tests.sh` to the scripts folder for consistency.

By submitting this pull request, I confirm that you can use, modify,
copy, and redistribute this contribution, under the terms of your
choice.