Copy benchmark-overhead from opentelemetry-java-instrumentation to performance-tests
In this commit we are copying all content from https://github.com/open-telemetry/opentelemetry-java-instrumentation/tree/main/benchmark-overhead to https://github.com/aws-observability/aws-otel-python-instrumentation/tree/main/performance-tests. We will be using benchmark-overhead as a template to building our own performance tests, so initially we are just copying all content over. Much of this content will be changed and/or deleted in upcoming commits, but for now, it's just a raw copy-paste.