Make spotless work
In this commit, we are making spotless work, as I found out that
spotless is not being run in the upstream for benchmark-overhead. To
achive this, I've added a license file to `scripts` and updated
build.gradle.kts to reference it. Then I ran spotlessApply to bring the
code up to expectations. In future commits, we will automatically run
spotless to ensure code is up to standards on every commit.