Rename PetClinic and Agent (#81)
In this commit, we are simply renaming PetClinic and Agent to
VehicleInventoryService and DistroConfig. We are doing this change so we
can have multiple people working on this in parallel without stepping on
toes.

Note that there are a few occurences of PetClinic/Agent that were not
changed - in these cases, we know that we will be deleting this logic,
so we did not bother updating now (e.g. AgentResolver).

Testing: 
```
dev-dsk-thp-2a-bd3a87ed % ./gradlew test

> Task :compileTestJava
Note: Some input files use or override a deprecated API.
Note: Recompile with -Xlint:deprecation for details.

Deprecated Gradle features were used in this build, making it incompatible with Gradle 9.0.

You can use '--warning-mode all' to show the individual deprecation warnings and determine if they come from your own scripts or plugins.

For more on this, please refer to https://docs.gradle.org/8.6/userguide/command_line_interface.html#sec:command_line_warnings in the Gradle documentation.

BUILD SUCCESSFUL in 5m 45s
3 actionable tasks: 3 executed
```

By submitting this pull request, I confirm that you can use, modify,
copy, and redistribute this contribution, under the terms of your
choice.