Update performance test to use python application and distro (#86)
In this commit we are removing the old logic for using a java
application and agent and adding in logic to use a python application
(VehicleInventory) and the ADOT distro. A few other relevant changes:
* All results (final and intermediary) now go to results folder, which
is added to gitignore
* Update postgres container to use v14.0 for application compatibility
* Commented out jfr-related code (heap stats, warm up, start recording)

By submitting this pull request, I confirm that you can use, modify,
copy, and redistribute this contribution, under the terms of your
choice.