Improve local operation resolution (#113)
*Issue:*
In the
[`get_ingress_operation()`](https://github.com/aws-observability/aws-otel-python-instrumentation/blob/main/aws-opentelemetry-distro/src/amazon/opentelemetry/distro/_aws_span_processing_util.py#L43)
method, we try to resolve the `aws.local.operation` by first using the
span name. If the span name is not meaningful, we [try to generate the
value](https://github.com/aws-observability/aws-otel-python-instrumentation/blob/main/aws-opentelemetry-distro/src/amazon/opentelemetry/distro/_aws_span_processing_util.py#L162)
by combining `http.method` and `http.target` values from the span. If
the `http.target` value is not present, we set the `aws.local.operation`
to `UnknownOperation`.

In our testing, we found that quite often we get ingress requests where
the span name is not meaningful and the `http.target` is also not set by
the instrumentation. So we end up getting a lot of `UnknownOperation`
for such requests.

*Proposal:*
If the span name is not meaningful, and the `http.target` is not
present, then we try and parse the `http.url` if present to only pick
the first part of the url path (to minimize the probability of high
cardinality values) and combine it with the `http.method` to get the
local operation.
Similar logic is employed in [resolving the remote
operation](https://github.com/aws-observability/aws-otel-python-instrumentation/blob/566d1629b3ce4d26436c6dcad52f55f2b9e5873a/aws-opentelemetry-distro/src/amazon/opentelemetry/distro/_aws_metric_attribute_generator.py#L302-L320).


By submitting this pull request, I confirm that you can use, modify,
copy, and redistribute this contribution, under the terms of your
choice.