Apply patched only when dependency detected. (#122)
*Issue #, if available:*
We face `botocore` version conflict and python module miss localization
when both sample app and aws distro install `botocore` pkg with their
own version.

*Description of changes:*
Move `botocore` to `[project.optional-dependencies]` as additional
packages. The user will need to manually install botocore themselves,
otherwise, it won’t be automatically installed when users install
`aws-opentelemetry-distro`. This has the same behaviour as upstream
[opentelemetry-instrumentation-botocore](https://github.com/open-telemetry/opentelemetry-python-contrib/tree/b84d779104b8c511ae7b876f5a9f0db3f21654e6/instrumentation/opentelemetry-instrumentation-botocore).

Note:
1. To check whether `botocore` is installed in the system, we don't read
`pyproject.toml` file directly since it is only used when building the
pkg, and won't be copied into the container. Instead, we define a
`patch_libraries` dict.
2. we move `import apply_instrumentation_patches ` into conditional
check since the `botocore` only available if condition pass. Otherwise,
if we put it on the top of the file, the instrumentation will throw
error when `botocore` not installed.

Test:
Tested on ECS and locally.

By submitting this pull request, I confirm that you can use, modify,
copy, and redistribute this contribution, under the terms of your
choice.