Fix pr_build (#130)
pr_build had a dependency on a repo secret, which would have made it
fail for forked repos. In this commit we are removing that dependency,
as well as a few other non-required fields.

By submitting this pull request, I confirm that you can use, modify,
copy, and redistribute this contribution, under the terms of your
choice.