Add image scanning (#136)
In this PR we are adding image scanning workflows. We are adding a daily
scanner that we will monitor and alarm on, as well as a scanner that can
be run per-PR/build/release. Further, we are adding post-release steps
to update the daily scanner to look at the latest released image.

Tested `sed` command used in `post_release_version_bump.yml` here:
https://github.com/aws-observability/aws-otel-python-instrumentation/actions/runs/8530266063/job/23367719655?pr=137

By submitting this pull request, I confirm that you can use, modify,
copy, and redistribute this contribution, under the terms of your
choice.