Pre/Post release workflow update for Safety in Patch release (#144)
*Description of changes:*
1. For post release version bump, if the released version is a older
major.minor version (e.g. patch for 1.0.0 -> 1.0.1, but main in
1.1.0.dev0), it won't trigger

https://github.com/XinRanZhAWS/aws-otel-python-instrumentation/actions/runs/8619606136
2. For pre release, update the step for checkout release/v branches,
make it safe for patch release (in this case the release/v branch
already exist), and test to make sure it consume version number (and
changes) from release/v not main in this case

https://github.com/XinRanZhAWS/aws-otel-python-instrumentation/actions/runs/8619568106

https://github.com/XinRanZhAWS/aws-otel-python-instrumentation/pull/9/files
  

By submitting this pull request, I confirm that you can use, modify,
copy, and redistribute this contribution, under the terms of your
choice.