Update release_build.yml to reorder image releases (#147)
*Issue #, if available:*

*Description of changes:*
Reordering the release steps in release_build.yml. 
- The old order is Public and Private ECR -> PyPI -> GitHub
- The new order is PyPI -> Public ECR -> GitHub -> Private ECR.

### Testing
Tested by creating a fork of this repo and running the release workflow.
Changed the repos to my own private/public ECR repos and the images were
published successfully as well as a draft release was created. One thing
to callout, I removed the PyPI step as this step wasn't changed and I
didn't see a reason to test publishing there.

By submitting this pull request, I confirm that you can use, modify,
copy, and redistribute this contribution, under the terms of your
choice.