Attempt to fix test_urls_excluded_from_sampling (#158)
`test_urls_excluded_from_sampling` keeps failing on PR, causing annoying
delays. I suspect that the problem is a race condition - we are
exporting a span then grabbing it, and I believe that the delay between
export and being "grabbable" is causing the test to fail (typically with
`1 != 2`). As an attempted fix, I'm adding some retry logic that will
attempt to grab for a full second.

By submitting this pull request, I confirm that you can use, modify,
copy, and redistribute this contribution, under the terms of your
choice.