Metric Schema changes (#150)
In this commit, we are removing RemoteTarget and replacing with
RemoteResourceIdentifier and RemoteResourceType. Further, we are
formatting RemoteService, and the content of the RemoteResource
attributes such that they align with AWS Cloud Control resource names.

By submitting this pull request, I confirm that my contribution is made
under the terms of the Apache 2.0 license.