Fail artifact build if load and push are both false (#195)
If both are false, we get somewhat arcane failures. All existing
workflows set one to true and the other to false, this just enforces it.

Also, Fix a codespell issue that was introduced in 2.3.0 (previously
using 2.2.6).

Testing: Set both options to false in pr_build-
https://github.com/aws-observability/aws-otel-python-instrumentation/actions/runs/9230011064/job/25397368739?pr=195

By submitting this pull request, I confirm that you can use, modify,
copy, and redistribute this contribution, under the terms of your
choice.