Clean up patch tests (#212)
Follow up from
https://github.com/aws-observability/aws-otel-python-instrumentation/pull/204,
which had to be discarded, but changes to
`test_instrumentation_patch.py` still have value. Class had to be
significantly re-written for reasons that are made clear in the
class-level comments.

By submitting this pull request, I confirm that you can use, modify,
copy, and redistribute this contribution, under the terms of your
choice.