Move to public ECR for DockerFile
In this commit, we are updating the Dockerfile to use python3.11 and rust1.75 images to build the ADOT Python Instrumentation image vended by this repo. This should be a no-op as a) the public ECRs should be mirrors of the DockerHub equivalent and b) these images are only used in the build process, not the final produced image (which is built from `scratch`).