Move to public ECR for DockerFile (#235)
In this commit, we are updating the Dockerfile to use python3.11 and
rust1.75 images to build the ADOT Python Instrumentation image vended by
this repo. This should be a no-op as a) the public ECRs should be
mirrors of the DockerHub equivalent and b) these images are only used in
the build process, not the final produced image (which is built from
`scratch`).

We will be using:
* https://gallery.ecr.aws/docker/library/python
* https://gallery.ecr.aws/docker/library/rust

By submitting this pull request, I confirm that you can use, modify,
copy, and redistribute this contribution, under the terms of your
choice.