Add Contract Tests for SNS (#296)
*Description of changes:*
Set up SNS contract tests cases for `AWS::SNS::Topic` resource.

Note: Was not able to get the `error` path to work so added a `TODO`
comment for now. It seems boto3 handles the response code for this
resource in a special way compared to the other resources. Will need to
investigate further to figure out how we can add this case.

*Test Plan:*

![Screenshot 2024-11-27 at 10 52
37 AM](https://github.com/user-attachments/assets/121a5829-e960-4292-b3c5-8ceb9c6f9d6c)

By submitting this pull request, I confirm that you can use, modify,
copy, and redistribute this contribution, under the terms of your
choice.