remove lambda remote service override configuration (#325)
**Issue #, if available:**
If user decides to override `RemoteService value`, the Lambda Topology
issue solution will not be able to support multiple downstream Lambdas.

For example, consider lambdaA calls both lambdaB and lambdaC. If the
user overrides `LAMBDA_APPLICATION_SIGNALS_REMOTE_SERVICE`, it will
effectively hardcode the `RemoteService` value. As a result, we are left
with an either/or situation where only one node will be connected to
lambdaA in the topology depending on the hardcoded value. This issue is
not present when the user does not override this `RemoteService` value
because that attribute will dynamically take on whatever the lambda
function name is.


My proposal is we remove the `LAMBDA_APPLICATION_SIGNALS_REMOTE_SERVICE`
env var. That is, we do not allow customers to override the
`RemoteService` attribute value. This should not cause issues for
customers to lose this configuration option. If the customer needs to
change the downstream lambda service name, they can change the lambda
function name itself rather than having to use an env var.

**Description of changes:**
Removing option to override `RemoteService` attribute value.

**Test plan:**
Sanity e2e test by building custom lambda layer and ensuring the correct
EMF Logs are still generated after change.

By submitting this pull request, I confirm that you can use, modify,
copy, and redistribute this contribution, under the terms of your
choice.