Bump OTel Dependency Versions to 1.33.0/0.54b0 (#388)
## What does this pull request do?
Bumps our OTel Dependency versions to
[1.33.0/0.54b0](https://github.com/open-telemetry/opentelemetry-python/releases/tag/v1.33.0)
to support compatability with third-party AI Instrumentation
libraries/frameworks such as OpenInference, Traceloop/Openllmetry, and
OpenLit.

We do not bump to the latest upstream version
[1.34.0/0.55b0](https://github.com/open-telemetry/opentelemetry-python/releases/tag/v1.34.0)
because that release includes `BatchLogRecordProcessor` refactoring
which is not compatible with our Caton changes.


By submitting this pull request, I confirm that you can use, modify,
copy, and redistribute this contribution, under the terms of your
choice.