[Genesis] Add support for Log paths for LLOs in Header (#401)
*Description of changes:*
- In order to support CW Logs implementation for LLO, they need a new
header for the log object that contains all of the paths to the String
leaf nodes in the log body represented in `JSON protobuf` format.

- Increased the depth traversal for the Logs as we have a new schema for
consolidated log event.
*Testing*
- Added additional unit tests for `AwsBatchLogRecordProcessor` to
support the new LLO handling logic
- Manual E2E testing to verify that the logic for batching and LLO path
header works for Large Logs (mocked the size of the log by setting the
log size to 1,000,000 bytes)

By submitting this pull request, I confirm that you can use, modify,
copy, and redistribute this contribution, under the terms of your
choice.