Bump OTel Dependency Versions to 1.33.1/0.54b1 (#399)
## What does this pull request do?

This reverts commit a538775495db1f6a9a914a29784329a0a46ebf44 to re-apply
the OTel dependency update in our ADOT SDK since the root cause of the
failing contract tests was fixed:
https://github.com/aws-observability/aws-otel-python-instrumentation/pull/398.

We also bump the OTel dependency version to sync with our ADOT SDK bump.
This solves the issue of our contract tests not catching version bump
issues in the PR build.


By submitting this pull request, I confirm that you can use, modify,
copy, and redistribute this contribution, under the terms of your
choice.