Fix log race condition (#407)
*Issue #, if available:*
It is an existing issue from upstream that only exists in certain cases
1. Application using [ASGI](https://asgi.readthedocs.io/en/latest/)
framework like FastAPI, it will always fork another thread from main
thread to handle the traffic
2. Application has tried to log something during the initialization
phase and failed.

This explains all our confusions thatWhy it happens only on FastAPI
application but not Flask in our test, because the latter one is not a
ASGI framework It happens on TraceLoop SDK but not in OpenLLMetry,
because TraceLoop will log a warning message during initialization but
the latter one does not

*Fixes:*
The race condition is cause because the default LogProcessor using
trying to export data with a blocking call, which blocked the worker
thread in ASGI framework. I am changing it to non-blocking call so it
will not run into race condition issue.

By submitting this pull request, I confirm that you can use, modify,
copy, and redistribute this contribution, under the terms of your
choice.