add LoggerProvider force flush in Lambda function (#441)
*Issue #, if available:*

The current Lambda instrumentation is missing a call to the
loggerProvider.force_flush method, which can result in delayed or
missing OTel logs in the Lambda environment due to Lambda freeze

*Description of changes:*
Adding LoggerProvider force flush in Lambda instrumentation.
Logging a debug log if the loggerProvider does not support force flush
because the default global LoggerProvider is ProxyLoggerProvider in
OpenTelemetry Python, will be replaced to SDKLoggerProvider only if user
set environment variable
`OTEL_PYTHON_LOGGING_AUTO_INSTRUMENTATION_ENABLED = true`, which is not
set in ADOT Lambda layer.

By submitting this pull request, I confirm that you can use, modify,
copy, and redistribute this contribution, under the terms of your
choice.

Co-authored-by: Jonathan Lee <107072447+jj22ee@users.noreply.github.com>