Post release 0.12.1: Update version to 0.12.1.dev0 (#471)
This PR is manually created to replace
https://github.com/aws-observability/aws-otel-python-instrumentation/pull/469

This PR prepares the main branch for the next development cycle by
updating the version to 0.12.1.dev0 and updating the image version to be
scanned to the latest released.

This PR should only be merge when release for version v0.12.1 is
success.

By submitting this pull request, I confirm that you can use, modify,
copy, and redistribute this contribution, under the terms of your
choice.