Upload source distribution to pypi (#532)
*Issue:* 
Generally a good practice to upload sdist along with the wheel artifact
to PyPI:
https://packaging.python.org/en/latest/guides/distributing-packages-using-setuptools/#source-distributions
OTel SDK package for reference:
https://pypi.org/project/opentelemetry-sdk/#files

*Description of changes:*
We are already building the source distribution tarball. This PR updates
the release workflow to just upload it.

*Testing:*
Tested a successful upload to test pypi in this modified release
workflow run (verified by @thpierce):
https://github.com/aws-observability/aws-otel-python-instrumentation/actions/runs/19081104321

By submitting this pull request, I confirm that you can use, modify,
copy, and redistribute this contribution, under the terms of your
choice.

---------

Co-authored-by: Thomas Pierce <thp@amazon.com>