Not patch code attributes if the input is not function or method (#538)
*Issue #, if available:*

When a Starlette user registers an ASGI app via the route registration
method, the ADOT Python SDK’s Starlette instrumentation mistakenly
treats it as a regular synchronous function. As a result, the patched
ASGI app cannot be properly registered in Starlette, preventing
Starlette from starting the app. This leads clients to receive 405
Method Not Allowed errors.

Since FastMCP registers ASGI apps with Starlette, this bug causes MPC
services built with FastMCP to fail to start.

*Description of changes:*
When Starlette’s route registration method receives a callable that is
neither a function nor a method, the ADOT Python SDK leaves it
unmodified, allowing the ASGI app to be registered and started
correctly.

By submitting this pull request, I confirm that you can use, modify,
copy, and redistribute this contribution, under the terms of your
choice.